tests: remove module-specific bits from test_utils
Test utils is a global library, but had lots of bits from modules (e.g.
security, model, etc...). This resulted in a lot of implicit
dependencies.

This commit splits it and includes the specializations in each module.
